+++ lib/swiss_knife.dart
@@ -3,6 +3,7 @@ library;
 export 'src/collections.dart';
 export 'src/data.dart';
 export 'src/date.dart';
+export 'src/dual_map.dart';
 export 'src/events.dart';
 export 'src/expando_with_finalizer.dart';
 export 'src/flat_hash_map.dart';